requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
Faker==19.13.0
openai>=1.0.0
python-dotenv==1.0.0
supabase>=1.0.3 